Article press_release template plus default article template

+++ helpers/fetch_article_type_from_yaml.js
@@ -10,6 +10,7 @@ const STRAPIDATA = yaml.safeLoad(fs.readFileSync(strapiDataPath, 'utf8'))
 const DOMAIN = process.env['DOMAIN'] || 'poff.ee'
 const STRAPIDIR = '/uploads/'
 const STRAPIHOSTWITHDIR = `http://${process.env['StrapiHost']}${STRAPIDIR}`;
+const DEFAULTTEMPLATENAME = 'news'
 
 const mapping = {
     'poff.ee': 'POFFiArticle',
@@ -146,7 +147,11 @@ function getDataCB(data, dirPath, lang, writeIndexFile, dataFrom, showErrors, ge
             let yamlStr = yaml.safeDump(element, { 'indent': '4' });
 
             fs.writeFileSync(`${element.directory}/data.${lang}.yaml`, yamlStr, 'utf8');
-            fs.writeFileSync(`${element.directory}/index.pug`, `include /_templates/article_${artType.name}_index_template.pug`)
+            if (fs.existsSync(`include /_templates/article_${artType.name}_index_template.pug`)) {
+                fs.writeFileSync(`${element.directory}/index.pug`, `include /_templates/article_${artType.name}_index_template.pug`)
+            } else {
+                fs.writeFileSync(`${element.directory}/index.pug`, `include /_templates/article_${DEFAULTTEMPLATENAME}_index_template.pug`)
+            }
         }
     }
 }
